+++ src/platform/plugins/shared/controls/public/actions/clear_control_action.tsx
@@ -11,15 +11,43 @@ import React, { SyntheticEvent } from 'react';
 
 import { EuiButtonIcon, EuiToolTip } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import type { EmbeddableApiContext, HasUniqueId } from '@kbn/presentation-publishing';
+import {
+  apiCanAccessViewMode,
+  apiHasParentApi,
+  apiHasType,
+  apiHasUniqueId,
+  apiIsOfType,
+  HasParentApi,
+  type EmbeddableApiContext,
+  type HasUniqueId,
+  HasType,
+} from '@kbn/presentation-publishing';
 import {
   IncompatibleActionError,
   FrequentCompatibilityChangeAction,
   type Action,
 } from '@kbn/ui-actions-plugin/public';
-import { isClearableControl } from '../types';
+import { PresentationContainer, apiIsPresentationContainer } from '@kbn/presentation-containers';
+import { CONTROL_GROUP_TYPE } from '../../common';
+import { CanClearSelections, isClearableControl } from '../types';
 
-import { ACTION_CLEAR_CONTROL } from '.';
+import { ACTION_CLEAR_CONTROL } from './constants';
+
+type ClearControlActionApi = HasType &
+  HasUniqueId &
+  CanClearSelections &
+  HasParentApi<PresentationContainer & HasType>;
+
+const compatibilityCheck = (api: unknown | null): api is ClearControlActionApi =>
+  Boolean(
+    apiHasType(api) &&
+      apiHasUniqueId(api) &&
+      isClearableControl(api) &&
+      apiHasParentApi(api) &&
+      apiCanAccessViewMode(api.parentApi) &&
+      apiIsOfType(api.parentApi, CONTROL_GROUP_TYPE) &&
+      apiIsPresentationContainer(api.parentApi)
+  );
 
 export class ClearControlAction
   implements Action<EmbeddableApiContext>, FrequentCompatibilityChangeAction<EmbeddableApiContext>
@@ -73,12 +101,10 @@ export class ClearControlAction
   }
 
   public async isCompatible({ embeddable }: EmbeddableApiContext) {
-    const { isCompatible } = await import('./clear_control_action_compatibility_check');
-    return isCompatible(embeddable);
+    return compatibilityCheck(embeddable);
   }
 
   public async execute({ embeddable }: EmbeddableApiContext) {
-    const { compatibilityCheck } = await import('./clear_control_action_compatibility_check');
     if (!compatibilityCheck(embeddable)) throw new IncompatibleActionError();
 
     embeddable.clearSelections();
