+++ src/platform/plugins/shared/controls/public/actions/delete_control_action.tsx
@@ -11,11 +11,39 @@ import React from 'react';
 
 import { EuiButtonIcon, EuiToolTip } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import type { HasUniqueId, EmbeddableApiContext } from '@kbn/presentation-publishing';
+import {
+  type HasUniqueId,
+  type EmbeddableApiContext,
+  type HasType,
+  type HasParentApi,
+  type PublishesViewMode,
+  apiHasType,
+  apiHasUniqueId,
+  apiHasParentApi,
+  apiCanAccessViewMode,
+  apiIsOfType,
+  getInheritedViewMode,
+} from '@kbn/presentation-publishing';
 import { IncompatibleActionError, type Action } from '@kbn/ui-actions-plugin/public';
 
-import { ACTION_DELETE_CONTROL } from '.';
+import { PresentationContainer, apiIsPresentationContainer } from '@kbn/presentation-containers';
+import { CONTROL_GROUP_TYPE } from '../../common';
+import { ACTION_DELETE_CONTROL } from './constants';
 import { coreServices } from '../services/kibana_services';
+
+type DeleteControlActionApi = HasType &
+  HasUniqueId &
+  HasParentApi<PresentationContainer & PublishesViewMode & HasType>;
+
+export const compatibilityCheck = (api: unknown | null): api is DeleteControlActionApi =>
+  Boolean(
+    apiHasType(api) &&
+      apiHasUniqueId(api) &&
+      apiHasParentApi(api) &&
+      apiCanAccessViewMode(api.parentApi) &&
+      apiIsOfType(api.parentApi, CONTROL_GROUP_TYPE) &&
+      apiIsPresentationContainer(api.parentApi)
+  );
 
 export class DeleteControlAction implements Action<EmbeddableApiContext> {
   public readonly type = ACTION_DELETE_CONTROL;
@@ -49,12 +77,10 @@ export class DeleteControlAction implements Action<EmbeddableApiContext> {
   }
 
   public async isCompatible({ embeddable }: EmbeddableApiContext) {
-    const { isCompatible } = await import('./delete_control_action_compatibility_check');
-    return isCompatible(embeddable);
+    return compatibilityCheck(embeddable) && getInheritedViewMode(embeddable.parentApi) === 'edit';
   }
 
   public async execute({ embeddable }: EmbeddableApiContext) {
-    const { compatibilityCheck } = await import('./delete_control_action_compatibility_check');
     if (!compatibilityCheck(embeddable)) throw new IncompatibleActionError();
 
     coreServices.overlays
