--- src/platform/plugins/shared/controls/public/actions/index.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-export const ACTION_EDIT_CONTROL = 'editDataControl';
-export const ACTION_CLEAR_CONTROL = 'clearControl';
-export const ACTION_DELETE_CONTROL = 'deleteControl';
