+++ src/platform/plugins/shared/controls/public/control_group/control_group_renderer/control_group_renderer_lazy.tsx
@@ -12,7 +12,7 @@ import { dynamic } from '@kbn/shared-ux-utility';
 import type { ControlGroupRendererProps } from './control_group_renderer';
 
 const Component = dynamic(async () => {
-  const { ControlGroupRenderer } = await import('./control_group_renderer');
+  const { ControlGroupRenderer } = await import('../../controls_module');
   return {
     default: ControlGroupRenderer,
   };
