+++ src/platform/plugins/shared/controls/public/controls_module.ts
@@ -1,0 +1,20 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+export { ClearControlAction } from './actions/clear_control_action';
+export { DeleteControlAction } from './actions/delete_control_action';
+export { EditControlAction } from './actions/edit_control_action';
+
+export { getControlGroupEmbeddableFactory } from './control_group/get_control_group_factory';
+
+export { getOptionsListControlFactory } from './controls/data_controls/options_list_control/get_options_list_control_factory';
+export { getRangesliderControlFactory } from './controls/data_controls/range_slider/get_range_slider_control_factory';
+export { getTimesliderControlFactory } from './controls/timeslider_control/get_timeslider_control_factory';
+
+export { ControlGroupRenderer } from './control_group/control_group_renderer/control_group_renderer';
