+++ src/platform/plugins/shared/embeddable/public/index.ts
@@ -25,10 +25,8 @@ export {
   isValueClickTriggerContext,
   MULTI_VALUE_CLICK_TRIGGER,
   panelBadgeTrigger,
-  panelHoverTrigger,
   panelNotificationTrigger,
   PANEL_BADGE_TRIGGER,
-  PANEL_HOVER_TRIGGER,
   PANEL_NOTIFICATION_TRIGGER,
   SELECT_RANGE_TRIGGER,
   VALUE_CLICK_TRIGGER,
