+++ src/platform/plugins/shared/embeddable/public/ui_actions/register_triggers.ts
@@ -16,12 +16,10 @@ import {
   selectRangeTrigger,
   valueClickTrigger,
   cellValueTrigger,
-  panelHoverTrigger,
 } from './triggers';
 
 export const registerTriggers = (uiActions: UiActionsSetup) => {
   uiActions.registerTrigger(contextMenuTrigger);
-  uiActions.registerTrigger(panelHoverTrigger);
   uiActions.registerTrigger(panelBadgeTrigger);
   uiActions.registerTrigger(panelNotificationTrigger);
   uiActions.registerTrigger(selectRangeTrigger);
