--- src/platform/plugins/shared/presentation_util/public/components/floating_actions/floating_actions.scss
@@ -1,23 +1,0 @@
-.presentationUtil__floatingActionsWrapper {
-  position: relative;
-
-  .presentationUtil__floatingActions {
-    opacity: 0;
-    visibility: hidden;
-    // slower transition on hover leave in case the user accidentally stops hover
-    transition: visibility $euiAnimSpeedSlow, opacity $euiAnimSpeedSlow;
-
-    position: absolute;
-    right: $euiSizeXS;
-    top: (-$euiSizeL);
-    z-index: $euiZLevel9; // the highest possible z-level
-  }
-
-  &:hover, &:focus-within {
-    .presentationUtil__floatingActions {
-      opacity: 1;
-      visibility: visible;
-      transition: visibility $euiAnimSpeedFast, opacity $euiAnimSpeedFast;
-    }
-  }
-}
