+++ src/platform/plugins/shared/presentation_util/tsconfig.json
@@ -15,7 +15,6 @@
   "kbn_references": [
     "@kbn/core",
     "@kbn/saved-objects-plugin",
-    "@kbn/embeddable-plugin",
     "@kbn/kibana-react-plugin",
     "@kbn/i18n",
     "@kbn/expressions-plugin",
@@ -34,9 +33,7 @@
     "@kbn/code-editor",
     "@kbn/calculate-width-from-char-count",
     "@kbn/field-utils",
-    "@kbn/presentation-publishing",
     "@kbn/core-ui-settings-browser",
-    "@kbn/presentation-panel-plugin",
   ],
   "exclude": ["target/**/*"]
 }
